requirements.txt
streamlit>=1.30.0
groq>=0.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
